improvement: split personal and pmb service
