--- apps/api/general/src/app/app.module.ts
@@ -1,12 +1,0 @@
-import { Module } from "@nestjs/common";
-
-import { AppController } from "./app.controller";
-import { AppService } from "./app.service";
-import { PrismaModule } from "@uninus/api/modules";
-
-@Module({
-  imports: [PrismaModule],
-  controllers: [AppController],
-  providers: [AppService],
-})
-export class AppModule {}
