+++ apps/api/general/src/app/master/index.ts
@@ -1,0 +1,1 @@
+export * from "./app.module";
