+++ apps/api/general/src/app/personal/app.module.ts
@@ -1,0 +1,10 @@
+import { Module } from "@nestjs/common";
+import { PersonalController } from "./app.controller";
+import { PersonalService } from "./app.service";
+
+@Module({
+  imports: [],
+  controllers: [PersonalController],
+  providers: [PersonalService],
+})
+export class PersonalModule {}
