+++ apps/api/general/src/app/pmb/app.module.ts
@@ -1,0 +1,10 @@
+import { Module } from "@nestjs/common";
+import { PMBController } from "./app.controller";
+import { PMBService } from "./app.service";
+
+@Module({
+  imports: [],
+  controllers: [PMBController],
+  providers: [PMBService],
+})
+export class PMBModule {}
